requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
alembic
python-dotenv
firebase-admin